Change various LOCK2 calls to use DS_LOCK2 instead, DS_LOCK2 uses TRY_LOCK semantics on the two mutexes instead of LOCK semantics, and then releases the lock and loops with a delay to repeatedly try again if one of the locks fails.
This fixes some instability issues we are seeing where RPC randomly locks up and stops responding.

Fix an issue with the 'unlock' callback of 'walletpassphrase' whereby it could temporarily freeze the main RPC thread. In most cases this would result in slow RPC but in extreme cases it could be so slow that the slowness would then cascade into other parts of the program (e.g. peers timing out) presenting as a frozen RPC that no longer responds.