Change the function signature of some of the PoW2 phase helpers.
Change CheckTransaction to do some PoW2 related transaction validation - add a new 'checkHeight' paramater that is needed to verify that the 'lock' is not violated.
Fix a bug in RouletteItem::operator<
Add some extra mutex locks around code that had potential for race conditions.
New functions 'ForceActivateChainStep' and 'ForceActivateChain' that behave similar to 'ActivateBestChain' except with a specific target block in mind and without firing events, without updating various other parts of program state etc. - with the idea that it can be used by the witnessing code on a temporary chain without permanently altering program state in any way.
Extract some of the logic from 'GetWitness' into a new helper 'getAllUnspentWitnessCoins' that we reuse elsewhere in the code - also improve this code in various ways (multiple bug and robustness fixes)
Multiple bug/robustness fixes to 'GetWitness' - as well as logic changes (ForceActivateChain) that will allow it to work on any block (not just the chainTip)
Various bug fixes relating to the difference between coinbase/pow2coinbase needing to be more explicit in some cases.