Properly fix Rewrite CConnman::PushMessage to avoid race and remove assert.
In the previous fix it could happen that multiple async write actions on a socket take place at the same time.
This fix is based on the previous model. Kept state is reduced by consuming from the message queue when starting async write and properly reacting to an empty queue instead of asserting.