Build skiplist after erasing of blocks is done (could crash)
If a block is erased from the block index that has a pskip pointing to it that pointer will now be invalid.
This is a rare case but could happen and is hard to detect. It can occur for example if the pskip is accessed during a GetAncestor that will exactly hit on the automatically created block index preceding the partial chain start (which would have been deleted in the cleanup of the block index after loading).