Use internal/change keychain to determine change (instead of address book lookup)
Correctly calculating the change is needed to compute a sensible amount for the mutations that are derived for transactions to self.
For sure there are no address book entries for those. In general there will be no address book entries. If if there were they would be lost when starting clean and doing a recovery.