Revert "android: integrate protobuf code generation from wallet.proto into Gradle build"
This reverts commit 3d27c9fe40c39f41bdff6e78f70440cb411198af.

This code was intentionally comitted and reverted to keep the changes required to have the build generate the code from proto files. Should we decide to go this route at some point it will come in handy. Though at this point backed out of this route as besides the protobuf handling additional work is needed to read old wallet files, decryption in particular.