android: avoid upgrading after removing wallet
If an upgraded wallet is later removed and the app is re-initialized it is not expected to be prompted with the upgrade again.
The upgrade is recorded using a marker file, the old wallet proto file is kept in place.
This makes it possible to re-attempt an upgrade by manually deleting the marker file or publishing an emergency fix which still has the data available.