ANDROID: Improve handling of external URI's, especially if wallet not yet created
Move URI handling code out of the WalletActivity and into a special URI handling activity
Make this activity transparent so that just the bottom fragment can be displayed
Implement logic in the SendCoinsFragment to call "finish()" on its hosting activity IFF the hosting activity wants this behaviour (which we do in this case)
Detect the "special" case of invoking a URI handler before the wallet is even created and just put a warning toast instead.
Refactor parts of the "UnityCore" observer and "IntroActivity" for better interaction with all of the above.