BUILD: ANDROID: Improve configuration options for scripts
Rename "ndk_definitions.sh" as "ndk_defnitions.conf" to better clarify its configuration role (that it only sets variables and doesn't actually execute anything).
Introduce "private.conf" (same as used in localisation scripts) to the android scripts as well
Load the "ndk_defnitions" at the top before private.conf - so that it can be overriden in private.conf
Introduce "private.conf.example" to help document the options that can be set in private.conf
Introduce a basic README for the "developer-tools" directory.