ANDROID: BUILD: specify explicitly which targets to build (use --targets <comma delimited list> or set NDK_TARGETS)
Limiting the targets needed during development speeds up the build cycle.
This way it is easy to build just a specific (set of) target(s) without having to make changes.
Previously this could be done by removing the unwanted target specification files, at the risk of accidentally committing those changes.
Note that this is not necessary for iOS as there you can just do "make" in the build directory of the target you need.