CORE: Implement index based coin cache/utxo etc.
Update utxo 'Coin' format to include the transaction index (within a block) of the coin.
Update txdb, coin view etc. to index on *both* hash as well as height/index/n combination so that coins can be references by either.
Unfortunately this means existing wallets must do a rescan - so we implement code to facilitate this.