Fix text color of some elements on macOS in dark mode (new SDK build)
When building with a recent macOS SDK (and a recent Qt) some elements will have white text (on white background) in dark mode.
Note that the official builds are done on an older SDK (and Qt) and therefore run in compatibility mode which will not switch the text color to white in dark mode.