CORE: Fix a bug with InvalidateBlock command when block not in chain.
When InvalidateBlock is called on a block not in the chain (e.g. possibly previously invalidated or in a fork etc.) the chain tip was still pointlessly being rewound all the way to the height of the invalidated block.
Fix this so that this is only done when the block forms part of the chain.