BUILD: Work around obscure macOS build issue.
For some reason when we use '__builtin_cpu_supports' we end up with '___cpu_model' being undefined.
The compiler/linker is meant to define this for us, and does if we make a trivial test case, but is either breaking in more advanced cases or (more likely) our build system is somehow breaking the behaviour.
For now we workaround by directly building in the '___cpu_model' function, taken directly from clang source, ourself with an include - but we should fix this.