Introduce bucket hash on outpoint
The COutPoint::getHash returned a "synthesised" hash if the tx hash is not available. However this fails in places where it is expected to be the tx hash.
Though a hash is needed in all cases for the std::unordered_maps kept in coins.h
A bucket hash is introduced for that purpose, which is the first step in fixing this problem.