Fix CGuldenWallet::IsMine to use actual tx hash
Tx are registered in the wallet using the actual tx hash, with introduction of the index based outpoints this IsMine would suddenly try to lookup using a "synthesised" hash, failing to find the tx and thus giving the wrong IsMine result.
This fixed approach is slow because it will need to do a disk lookup (slow always beats incorrect), so this might be just an interim solution and may be replaced by a faster approach.
rebase into the GetTxHash function