Avoid using hash of COutPoint for index outpoints
The COutPoint hash was used as transaction hash which it wasn't for index based outpoints. The outpoint hash method now explicitly is named to make it clear that it refers to a tx hash, and is only valid in that case.
All cases where the outpoint hash was used has been  fixed/replaced to account for index based outpoints.