BUILD: Work around an issue with gitian builds
The changes for android/iOS compiles are so mehow breaking gitian builds (using the wrong nm).
It isn't clear to me why, so for now for the sake of pragmatism I just introduce a way to disable the changes for gitian, we should later figure this out and fix it properly though.