CORE: Fix issue with witnessing and index based outpoints
Witness sync has exposed an issue where witnessing was setting the prevout in a way that could cause issues (with index based outpoints); This would not actually be exposed in the wild as prior to witness sync all witness outpoints must be hash based.
Force use of getter/setter instead to ensure there are no other similar issues (there weren't) and that it can't happen again in any other new code.