MISC: Add height to various witness functions
We don't currently use this paramater but it brings the code more inline with Florin where it is used which makes merges easier.
And if we do ever want to make a similar height based change to weight formula as Florin has done it will be useful to already have the necessary changes in place.