CORE: Use chainparams for reg/testnet conditional checks
We were calling IsArgSet("testnet") etc. in a lot of places, IsArgSet is protected by cs_args which was creating a lot of lock contention.
As this is something that changes value once only (at startup) theres really no reason for the lock contention so we work around this by instead just having a single constant place this can be looked up (without locking) via the chainparams